break up core card file
