+++ frontend/packages/core/src/Cards/card.tsx
@@ -5,15 +5,12 @@ import type {
   CardHeaderProps as MuiCardHeaderProps,
 } from "@material-ui/core";
 import {
-  Avatar,
   Card as MuiCard,
-  CardActionArea,
-  CardActionAreaProps,
   CardContent as MuiCardContent,
   CardHeader as MuiCardHeader,
 } from "@material-ui/core";
 
-import { StyledTypography } from "./typography";
+import { StyledTypography } from "../typography";
 
 const StyledCard = styled(MuiCard)({
   boxShadow: "0px 4px 6px rgba(53, 72, 212, 0.2)",
@@ -65,54 +62,6 @@ interface CardContentProps extends MuiCardContentProps {}
 
 const CardContent = ({ children, ...props }: CardContentProps) => (
   <StyledCardContent {...props}>{children}</StyledCardContent>
-);
-
-const StyledLandingCard = styled(Card)({
-  border: "none",
-
-  "& .header": {
-    display: "inline-flex",
-    marginBottom: "16px",
-    fontWeight: "bold",
-    fontSize: "12px",
-    lineHeight: "36px",
-    color: "rgba(13, 16, 48, 0.6)",
-  },
-
-  "& .header .icon .MuiAvatar-root": {
-    height: "36px",
-    width: "36px",
-    marginRight: "8px",
-    color: "rgba(13, 16, 48, 0.38)",
-    backgroundColor: "rgba(13, 16, 48, 0.12)",
-  },
-});
-
-export interface LandingCardProps extends Pick<CardActionAreaProps, "onClick"> {
-  group: string;
-  title: string;
-  description: string;
-}
-
-export const LandingCard = ({ group, title, description, onClick, ...props }: LandingCardProps) => (
-  <StyledLandingCard {...props}>
-    <CardActionArea onClick={onClick}>
-      <CardContent>
-        <div className="header">
-          <div className="icon">
-            <Avatar>{group.charAt(0)}</Avatar>
-          </div>
-          <span>{group}</span>
-        </div>
-        <div>
-          <StyledTypography variant="h3">{title}</StyledTypography>
-          <StyledTypography style={{ color: "rgba(13, 16, 48, 0.6)" }} variant="body2">
-            {description}
-          </StyledTypography>
-        </div>
-      </CardContent>
-    </CardActionArea>
-  </StyledLandingCard>
 );
 
 export { Card, CardContent, CardHeader };
