+++ frontend/packages/core/src/Cards/stories/landing-card.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import type { Meta } from "@storybook/react";
 
-import type { LandingCardProps } from "../card";
-import { LandingCard } from "../card";
+import type { LandingCardProps } from "../landing-card";
+import { LandingCard } from "../landing-card";
 
 export default {
   title: "Core/Card/Landing Card",
