+++ frontend/packages/core/src/index.tsx
@@ -1,6 +1,7 @@
 import { Grid } from "@material-ui/core";
 
 import { userId } from "./AppLayout/user";
+import { Card, CardContent, CardHeader } from "./Cards/card";
 import { Checkbox, CheckboxPanel } from "./Input/checkbox";
 import { Form, FormRow } from "./Input/form";
 import Radio from "./Input/radio";
@@ -17,7 +18,6 @@ import {
 } from "./accordion";
 import ClutchApp from "./AppProvider";
 import { Button, ButtonGroup, ClipboardButton, IconButton } from "./button";
-import { Card, CardContent, CardHeader } from "./card";
 import { Chip } from "./chip";
 import Confirmation from "./confirmation";
 import { useWizardContext, WizardContext } from "./Contexts";
